[GC] Use unsigned int to track external memory
This CL checks external memory doesn't go negative, and uses
unsigned int.

Bug: 42203776
Reviewed-on: https://chromium-review.googlesource.com/c/v8/v8/+/5872235
Commit-Queue: Etienne Pierre-Doray <etiennep@chromium.org>
Cr-Commit-Position: refs/heads/main@{#96350}

+++ include/v8-internal.h
@@ -1049,7 +1049,7 @@ class Internals {
 
   // Soft limit for AdjustAmountofExternalAllocatedMemory. Trigger an
   // incremental GC once the external memory reaches this limit.
-  static constexpr int kExternalAllocationSoftLimit = 64 * 1024 * 1024;
+  static constexpr size_t kExternalAllocationSoftLimit = 64 * 1024 * 1024;
 
 #ifdef V8_MAP_PACKING
   static const uintptr_t kMapWordMetadataMask = 0xffffULL << 48;
